+++ src/Tests/FShade.GLSL.Tests/SimpleTests.fs
@@ -172,9 +172,9 @@ let ``PointSize shader``() =
 
 
 [<ReflectedDefinition>]
-let fillArray (array : Arr<N<10>, V4f>) denom =
+let fillArray (array : Arr<N<10>, V4f> ref) denom =
     for i in 0 .. 9 do
-        array.[i] <- V4f(float32 i / denom)
+        array.Value.[i] <- V4f(float32 i / denom)
 
 
 [<Test>]
@@ -185,14 +185,14 @@ let ``Fill Array with Function``() =
         fragment {
             let array = Arr<N<10>, V4f>()
 
-            10.0f |> fillArray array
+            10.0f |> fillArray &&array
 
             return array.[uniform?color]
         }
 
     let expectedFillArrayGLSL =
         [
-            "void .*_fillArray_.*\(vec4 array\[10], float denom\)"
+            "void .*_fillArray_.*\(inout vec4 array\[10], float denom\)"
             "{"
             "    for\(int i = 0; \(i < 10\); i\+\+\)"
             "    {"
@@ -1353,79 +1353,6 @@ let ``Tuple function arguments``() =
         }
 
     GLSL.shouldCompile [ Effect.ofFunction fs ]
-
-[<Struct; ReflectedDefinition>]
-type Vector2f =
-    val mutable X : float32
-    val mutable Y : float32
-    new (x, y)   = { X = x; Y = y }
-    new (v)      = { X = v; Y = v }
-    new (v: int) = Vector2f(float32 v)
-    new (v: V2f) = Vector2f(v.X, v.Y)
-
-[<ReflectedDefinition>]
-type CVector2f =
-    val mutable X : float32
-    val mutable Y : float32
-    new ()       = { X = 0.0f; Y = 0.0f }
-    new (x, y)   = { X = x; Y = y }
-    new (v)      = { X = v; Y = v }
-    new (v: int) = CVector2f(float32 v)
-    new (v: V2f) = CVector2f(v.X, v.Y)
-
-[<ReflectedDefinition>]
-let vectorDot2 (a: Vector2f) (b: Vector2f) =
-    a.X * b.X + a.Y * b.Y
-
-[<ReflectedDefinition>]
-let cvectorDot2 (a: CVector2f) (b: CVector2f) =
-    a.X * b.X + a.Y * b.Y
-
-[<Test>]
-let ``Custom struct with constructors``() =
-    Setup.Run()
-
-    let fs (v : Vertex) =
-        fragment {
-            let a = Vector2f()
-            let b = Vector2f(v.hugo.XY)
-            let mutable c = Vector2f()
-            c <- Vector2f v.foo.X
-            c.X <- v.foo.Z
-            let d = Vector2f(3)
-            return a.X + a.Y + vectorDot2 b c + d.X
-        }
-
-    GLSL.shouldCompileAndContainRegex [ Effect.ofFunction fs ] [
-        "Vector2f a;"
-        "Vector2f_of_float32\("
-        "Vector2f_of_float32_float32\("
-        @"Vector2f_of_Aardvark_Base_V2f\(fs_hugo\.xy\)"
-        @"\+ new_SingleEffects_Vector2f_of_int32\(3\)\.X"
-    ]
-
-[<Test>]
-let ``Custom class with constructors``() =
-    Setup.Run()
-
-    let fs (v : Vertex) =
-        fragment {
-            let a = CVector2f()
-            let b = CVector2f(v.hugo.XY)
-            let mutable c = CVector2f()
-            c <- CVector2f v.foo.X
-            c.X <- v.foo.Z
-            let d = CVector2f(3)
-            return a.X + a.Y + cvectorDot2 b c + d.X
-        }
-
-    GLSL.shouldCompileAndContainRegex [ Effect.ofFunction fs ] [
-        "CVector2f a = new_SingleEffects_CVector2f\(\);"
-        "CVector2f_of_float32\("
-        "CVector2f_of_float32_float32\("
-        @"CVector2f_of_Aardvark_Base_V2f\(fs_hugo\.xy\)"
-        @"\+ new_SingleEffects_CVector2f_of_int32\(3\)\.X"
-    ]
 
 let boundedSampler1 =
     sampler2d {